esp32 change

+++ esp32/src/head_control.h
@@ -3,6 +3,7 @@
 
 #include <Arduino.h>
 
+
 // ============================================================
 // 初期化・定期更新
 // ============================================================
@@ -11,41 +12,37 @@
 void initHeadControl();
 
 // loop()から毎回呼ぶ。
-// delay()を使わず、現在角度を目標角度へ徐々に近づける。
+// 現在角度を目標角度へ1度ずつ近づける。
 void updateHeadControl();
 
 
 // ============================================================
-// コマンド
+// 頭の操作
 // ============================================================
 
-// 1回呼ばれるたび、目標角度を少しだけ左へ動かす。
+// 1回呼ぶごとに目標角度を1度変更する。
 void headLeft();
-
-// 1回呼ばれるたび、目標角度を少しだけ右へ動かす。
 void headRight();
-
-// 1回呼ばれるたび、目標角度を少しだけ上へ動かす。
 void headUp();
-
-// 1回呼ばれるたび、目標角度を少しだけ下へ動かす。
 void headDown();
 
 // 目標角度を中央へ設定する。
-// 実際のサーボはupdateHeadControl()によって徐々に中央へ戻る。
+// 実際のサーボはupdateHeadControl()で1度ずつ中央へ戻る。
 void headCenter();
 
 
 // ============================================================
 // 状態取得
+// 戻り値をfloatで統一する。
 // ============================================================
 
-int getCurrentHeadYaw();
-int getCurrentHeadPitch();
+float getCurrentHeadYaw();
+float getCurrentHeadPitch();
 
-int getTargetHeadYaw();
-int getTargetHeadPitch();
+float getTargetHeadYaw();
+float getTargetHeadPitch();
 
 bool isHeadMoving();
+
 
 #endif
